--- src/components/page/about-us-page.tsx
@@ -1,3 +1,0 @@
-export default function AboutUsPage() {
-  return <h1>About Us Page</h1>;
-}
