--- src/components/page/basket-page.tsx
@@ -1,3 +1,0 @@
-export default function BasketPage() {
-  return <h1>Basket Page</h1>;
-}
