--- src/components/page/catalog-page.tsx
@@ -1,3 +1,0 @@
-export default function CatalogPage() {
-  return <h1>Catalog Product Page</h1>;
-}
