--- src/components/page/main-page.tsx
@@ -1,3 +1,0 @@
-export default function MainPage() {
-  return <h1>Main Page</h1>;
-}
