--- src/components/page/user-profile-page.tsx
@@ -1,3 +1,0 @@
-export default function UserProfilePage() {
-  return <h1>User Profile Page</h1>;
-}
