+++ src/feature/header/NavBar.tsx
@@ -14,7 +14,7 @@ export default function Navbar() {
           <Link to="/login">Login</Link>
         </li>
         <li>
-          <Link to="/register">Register</Link>
+          <Link to="/registration">Registration</Link>
         </li>
         <li>
           <Link to="/catalog">Catalog</Link>
