+++ src/feature/header/header.tsx
@@ -1,18 +1,19 @@
 import { Heart, ShoppingCart } from 'lucide-react';
 import { NavLink } from 'react-router';
 
+import { ModeToggle } from '@/feature/header/mode-toggle';
+
 import Currency from './currency';
-import Language from './language';
 import { CustomNavigationMenu } from './nav-menu';
 import { ProfileDropdownMenu } from './profile-dropdown';
 import Search from './search';
 
 export function Header() {
   return (
-    <header className="grid grid-cols-[1fr_150px_1fr] items-center bg-white p-4 shadow-xl">
+    <header className="bg-background grid grid-cols-[1fr_150px_1fr] items-center p-4 shadow-xl">
       <CustomNavigationMenu />
 
-      <div className="flex w-24 items-center justify-center">
+      <div className="flex w-24 items-center justify-center rounded-md bg-white">
         <img src="logo.svg" alt="logo" />
       </div>
 
@@ -20,7 +21,7 @@ export function Header() {
         <ul className="flex flex-row items-center gap-4">
           <Search />
           <Currency />
-          <Language />
+          <ModeToggle />
           <div className="flex gap-4">
             <NavLink to="/wishlist" className="text-xl font-normal tracking-wider">
               <Heart strokeWidth={1.5} size={32} />
