+++ src/pages/login-page.tsx
@@ -1,10 +1,32 @@
-import { Baby } from 'lucide-react';
+import { useContext, useEffect } from 'react';
+import { useNavigate, Link } from 'react-router';
+import { toast } from 'sonner';
+
+import AuthContext from '@/feature/auth/login/auth-provider';
+import { LoginForm } from '@/feature/auth/login/login-form';
 
 export default function LoginPage() {
+  const { AUTH_STATUS_KEY } = useContext(AuthContext);
+  const navigate = useNavigate();
+
+  useEffect(() => {
+    if (AUTH_STATUS_KEY) {
+      void navigate('/', { replace: true });
+      toast.info('You are already login!');
+    }
+  }, [AUTH_STATUS_KEY, navigate]);
+
   return (
-    <div className="mt-4 flex flex-col items-center justify-center">
-      <h1>Login Page</h1>
-      <Baby color="green" size={60} />
+    <div className="bg-muted flex min-h-svh flex-col items-center justify-center gap-6 p-6 md:p-10">
+      <div className="flex w-full max-w-sm flex-col gap-6">
+        <Link to="/" className="flex items-center gap-2 self-center font-medium">
+          <div className="flex h-10 w-10 items-center justify-center rounded-md bg-white">
+            <img src="logo.svg" alt="logo" />
+          </div>
+          Poster Shop
+        </Link>
+        <LoginForm />
+      </div>
     </div>
   );
 }
