+++ src/pages/registration-page.tsx
@@ -1,10 +1,25 @@
-import { UsersRound } from 'lucide-react';
+import { Link } from 'react-router';
+
+import { RegistrationForm } from '@/feature/auth/registration/registration-form';
+import type { RegistrationFormData } from '@/feature/auth/registration/types';
 
 export default function RegistrationPage() {
+  const handleRegister = async (data: RegistrationFormData) => {
+    await new Promise((res) => setTimeout(res, 100));
+    console.log(data);
+  };
+
   return (
-    <div className="mt-4 flex flex-col items-center justify-center">
-      <h1>Registration Page</h1>
-      <UsersRound color="red" size={60} data-testid="users-icon" />
+    <div className="bg-muted flex min-h-svh items-center justify-center p-6">
+      <div className="flex w-full max-w-xl min-w-xs flex-col gap-6">
+        <Link to="/" className="flex items-center gap-2 self-center font-medium">
+          <div className="flex h-10 w-10 items-center justify-center rounded-md bg-white">
+            <img src="logo.svg" alt="logo" />
+          </div>
+          Poster Store
+        </Link>
+        <RegistrationForm onRegister={(data) => handleRegister(data)} />
+      </div>
     </div>
   );
 }
